+++ pkg/sass-parser/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Add support for parsing selector expressions.
 
+* Add support for parsing the `supports()` function in `@import` modifiers.
+
 ## 0.4.15
 
 * Add support for parsing list expressions.